# internal/docs/swagger.yaml
basePath: /api
definitions:
  github_com_projecthelena_warden_internal_db.APIKey:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      keyPrefix:
        type: string
      lastUsed:
        type: string
      name:
        type: string
    type: object
...
  github_com_projecthelena_warden_internal_db.Incident:
    properties:
      affectedGroups:
        description: JSON array
        type: string
      createdAt:
        type: string
      description:
        type: string
      endTime:
        type: string
      id:
        type: string
      severity:
        description: minor | major | critical
        type: string
      startTime:
        type: string
      status:
        description: investigation | identified | ... | scheduled | in_progress |
          completed
        type: string
      title:
        type: string
      type:
        description: incident | maintenance
        type: string
    type: object
...
info:
  contact: {}
  description: Self-hosted uptime monitoring API by Project Helena.
  title: Warden API
  version: "1.0"
paths:
  /api-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              keys:
                items:
                  $ref: '#/definitions/github_com_projecthelena_warden_internal_db.APIKey'
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      parameters:
      - description: Key name
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              key:
                type: string
              message:
                type: string
            type: object
        "400":
          description: Name is required
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - api-keys
  /api-keys/{id}:
    delete:
      parameters:
      - description: Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - api-keys
...
  /incidents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                affectedGroups:
                  items:
                    type: string
                  type: array
                createdAt:
                  type: string
                description:
                  type: string
                endTime:
                  type: string
                id:
                  type: string
                severity:
                  type: string
                startTime:
                  type: string
                status:
                  type: string
                title:
                  type: string
                type:
                  type: string
              type: object
            type: array
        "500":
          description: Failed to fetch incidents
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List incidents
      tags:
      - incidents
    post:
      consumes:
      - application/json
      parameters:
      - description: Incident payload
        in: body
        name: body
        required: true
        schema:
          properties:
            affectedGroups:
              items:
                type: string
              type: array
            description:
              type: string
            severity:
              type: string
            startTime:
              type: string
            status:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_projecthelena_warden_internal_db.Incident'
        "400":
          description: Invalid request body
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create incident
      tags:
      - incidents
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer sk_live_..." — create keys in Settings > API Keys
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"